requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
PyPDF2==3.0.1
langchain-core>=0.1.0
langchain-openai>=0.0.1
langchain-groq>=0.0.1
pydantic>=2.0.0
neo4j>=5.13.0
python-dotenv>=1.0.0
